pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axioms-drf-py"
dynamic = ["version"]
description = "OAuth2/OIDC authentication and authorization for Django REST Framework APIs"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Abhishek Tiwari", email = "abhishek@abhishek-tiwari.com" }
]
keywords = [
    "django",
    "djangorestframework",
    "drf",
    "oauth2",
    "oidc",
    "jwt",
    "authentication",
    "authorization",
    "axioms",
    "security",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
]
requires-python = ">=3.8"
dependencies = [
    "django>=3.2",
    "djangorestframework>=3.12.0",
    "pyjwt>=2.8.0",
    "jwcrypto>=1.5.0",
    "python-box>=7.0.0",
    "requests>=2.25.0,<3",
    "django-environ>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "build>=1.0.0",
    "twine>=4.0.0",
    "sphinx>=7.0.0",
    "sphinx-book-theme>=1.0.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-autobuild>=2021.3.14",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-book-theme>=1.0.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-autobuild>=2021.3.14",
]

[project.urls]
Homepage = "https://github.com/abhishektiwari/axioms-drf-py"
Documentation = "https://github.com/abhishektiwari/axioms-drf-py"
Repository = "https://github.com/abhishektiwari/axioms-drf-py"
Issues = "https://github.com/abhishektiwari/axioms-drf-py/issues"
Changelog = "https://github.com/abhishektiwari/axioms-drf-py/releases"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/axioms_drf/_version.py"
version_scheme = "no-guess-dev"
local_scheme = "no-local-version"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
DJANGO_SETTINGS_MODULE = "tests.settings"
django_find_project = false
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
